backend/requirements.txt
flask
flask-cors
numpy
pillow
tensorflow==2.15.0
